+++ apps/commerce-api/src/main/java/com/loopers/application/required/OrderItemRepository.java
@@ -1,6 +1,6 @@
 package com.loopers.application.required;
 
-import com.loopers.domain.orderitem.OrderItem;
+import com.loopers.domain.order.orderitem.OrderItem;
 
 public interface OrderItemRepository {
     OrderItem save(OrderItem orderItem);
