+++ apps/commerce-api/src/main/java/com/loopers/domain/order/orderitem/OrderItem.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.orderitem;
+package com.loopers.domain.order.orderitem;
 
 import java.math.BigDecimal;
