+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/orderitem/OrderItemJpaRepository.java
@@ -2,7 +2,7 @@ package com.loopers.infrastructure.orderitem;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.loopers.domain.orderitem.OrderItem;
+import com.loopers.domain.order.orderitem.OrderItem;
 
 public interface OrderItemJpaRepository extends JpaRepository<OrderItem, Long> {
 }
