+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/orderitem/OrderItemRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.loopers.infrastructure.orderitem;
 import org.springframework.stereotype.Repository;
 
 import com.loopers.application.required.OrderItemRepository;
-import com.loopers.domain.orderitem.OrderItem;
+import com.loopers.domain.order.orderitem.OrderItem;
 
 import lombok.RequiredArgsConstructor;
